# example/swagger.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "example",
        "version": "1"
    },
    "components": {
        "schemas": {
            "NotAuthorized": {
                "type": "object",
                "properties": {
                    "message": {
                        "type": "string"
                    }
                }
            },
            "NotFound": {
                "type": "object",
                "properties": {
                    "message": {
                        "type": "string"
                    }
                }
            },
            "User": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "age": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "height": {
                        "type": "number",
                        "format": "float"
                    },
                    "labels": {
                        "type": "object"
                    }
                }
            },
            "PutUserRequest": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "age": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "height": {
                        "type": "number",
                        "format": "float"
                    },
                    "labels": {
                        "type": "object"
                    }
                }
            },
            "PatchUserRequest": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "age": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "height": {
                        "type": "number",
                        "format": "float"
                    },
                    "labels": {
                        "type": "object"
                    }
                }
            },
            "Post": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "title": {
                        "type": "string"
                    },
                    "body": {
                        "type": "string",
                        "format": "base64"
                    },
                    "draft": {
                        "type": "boolean"
                    },
                    "created": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
            "PutPostRequest": {
                "type": "object",
                "properties": {
                    "title": {
                        "type": "string"
                    },
                    "body": {
                        "type": "string",
                        "format": "base64"
                    },
                    "draft": {
                        "type": "boolean"
                    },
                    "created": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
            "PatchPostRequest": {
                "type": "object",
                "properties": {
                    "title": {
                        "type": "string"
                    },
                    "body": {
                        "type": "string",
                        "format": "base64"
                    },
                    "draft": {
                        "type": "boolean"
                    },
                    "created": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            }
        }
    },
    "paths": {
        "/users/{id}": {
            "get": {
                "operationId": "GetUser",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id from path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "GetUser operation 200 response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    }
                }
            },
            "patch": {
                "operationId": "PatchUser",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id from path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "PatchUser operation request body",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PatchUserRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "PatchUser operation 200 response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "operationId": "DeleteUser",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id from path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "DeleteUser operation 200 response"
                    }
                }
            }
        },
        "/users": {
            "get": {
                "operationId": "GetUsers",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "limit from query",
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "GetUsers operation 200 response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/User"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "operationId": "PutUser",
                "requestBody": {
                    "description": "PutUser operation request body",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PutUserRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "PutUser operation 200 response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/posts/{id}": {
            "get": {
                "operationId": "GetPost",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id from path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "GetPost operation 200 response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Post"
                                }
                            }
                        }
                    }
                }
            },
            "patch": {
                "operationId": "PatchPost",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id from path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "PatchPost operation request body",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PatchPostRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "PatchPost operation 200 response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Post"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "operationId": "DeletePost",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id from path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "DeletePost operation 200 response"
                    }
                }
            }
        },
        "/posts": {
            "get": {
                "operationId": "GetPosts",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "limit from query",
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "GetPosts operation 200 response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Post"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "operationId": "PutPost",
                "requestBody": {
                    "description": "PutPost operation request body",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PutPostRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "PutPost operation 200 response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Post"
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}